fix : 호출 최적화 위한 정보 넘겨주기 전역 state 구현
