+++ src/1_Page/Championship/ui/InfoHeader/index.tsx
@@ -1,4 +1,3 @@
-import { useNavigate } from "react-router-dom";
 import EndChampionshipPanel from "./ui/EndChampionshipPanel";
 // state
 import useToggleState from "../../../../4_Shared/model/useToggleState";
@@ -8,16 +7,15 @@ import { championshipTypes } from "../../../../4_Shared/constant/championshipTyp
 import { formatDateKoreanDate } from "../../../../4_Shared/lib/dateFormatter";
 import defaultTrophyImg from "../../../../4_Shared/assets/svg/rank.svg";
 import { useMyCommunityRoleIdx } from "../../../../4_Shared/lib/useMyInfo";
-import useEditChampionshipStore from "../../../../4_Shared/zustand/useEditChampionshipStore";
 
 const InfoHeader = (props: InfoHeaderProps) => {
   const { championshipInfo } = props;
-  const navigate = useNavigate();
+
   const isChampionshipEnd = championshipInfo.common_status_idx === 4;
   // 어드민
   const [community_role_idx] = useMyCommunityRoleIdx();
   const isAdmin = community_role_idx === 0;
-  const { setChampionshipInfo } = useEditChampionshipStore();
+
   const [isHeaderCollapsed, toggleHeader] = useToggleState();
   const [isEndModalOpen, toggleEndModal] = useToggleState(true);
 
@@ -89,17 +87,7 @@ const InfoHeader = (props: InfoHeaderProps) => {
               </p>
               {isAdmin && !isChampionshipEnd && (
                 <div className="flex gap-2 items-center justify-center">
-                  <button
-                    className="px-3 py-1 text-sm border border-current rounded-md hover:bg-white/10 transition-colors"
-                    onClick={() => {
-                      setChampionshipInfo(championshipInfo);
-                      navigate(
-                        `/championship-edit/edit/${championshipInfo.community_list_idx}`
-                      );
-                    }}>
-                    수정
-                  </button>
-                  <EndChampionshipPanel />
+                  <EndChampionshipPanel championshipInfo={championshipInfo} />
                 </div>
               )}
             </div>
