+++ src/1_Page/Championship/ui/InfoHeader/ui/EndChampionshipPanel/index.tsx
@@ -1,15 +1,34 @@
 import React from "react";
 import useToggleState from "../../../../../../4_Shared/model/useToggleState";
 import EndChampionshipModal from "./ui/EndChampionshipModal";
+import useEditChampionshipStore from "../../../../../../4_Shared/zustand/useEditChampionshipStore";
+import useParamInteger from "../../../../../../4_Shared/model/useParamInteger";
+import { useNavigate } from "react-router-dom";
 
-const EndChampionshipPanel = () => {
+const EndChampionshipPanel = (props: EndChampionshipPanelProps) => {
+  const { championshipInfo } = props;
+  const navigate = useNavigate();
+  const championshipListIdx = useParamInteger("championshipIdx");
   const [isModalOpen, handleToggleModal] = useToggleState();
+  const { setChampionshipInfo, setChampionshipListIdx } =
+    useEditChampionshipStore();
 
   const cachedChampionshipEndDataRef = React.useRef<ChampionshipEndData>(
     {} as ChampionshipEndData
   );
   return (
     <div>
+      <button
+        className="px-3 py-1 text-sm border border-current rounded-md hover:bg-white/10 transition-colors"
+        onClick={() => {
+          setChampionshipListIdx(championshipListIdx);
+          setChampionshipInfo(championshipInfo);
+          navigate(
+            `/championship-edit/edit/${championshipInfo.community_list_idx}`
+          );
+        }}>
+        수정
+      </button>
       <button
         onClick={handleToggleModal}
         className="px-3 py-1 text-sm border border-current rounded-md hover:bg-white/10 transition-colors">
