+++ src/1_Page/Championship/ui/InfoHeader/ui/EndChampionshipPanel/type.d.ts
@@ -1,0 +1,3 @@
+type EndChampionshipPanelProps = {
+  championshipInfo: ChampionshipInfo;
+};
